review-like-api: fix
