+++ app/Actions/Dislike/DislikeReviewResponse.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Hedonist\Actions\Dislike;
+
+class DislikeReviewResponse 
+{
+
+}
