+++ app/Actions/Like/LikeReviewAction.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Hedonist\Actions\Review\Dislike;
+namespace Hedonist\Actions\Like;
 
 use Hedonist\Repositories\Like\{LikeRepository,LikeReviewCriteria};
 use Hedonist\Repositories\Dislike\{DislikeRepository,DislikeReviewCriteria};
 use Hedonist\Repositories\Review\ReviewRepository;
 use Hedonist\Entities\Review\Review;
-use Hedonist\Entities\Dislike\Dislike;
+use Hedonist\Entities\Like\Like;
 use Illuminate\Support\Facades\Auth;
 
 class LikeReviewAction
@@ -20,7 +20,7 @@ class LikeReviewAction
         $this->dislikeRepository = $dislikeRepository;
     }
 
-    public function execute(DislikeReviewRequest $request): DislikeReviewResponse
+    public function execute(LikeReviewRequest $request): LikeReviewResponse
     {        
         $reviewId = $request->getReviewId();
         $userId = Auth::id();
@@ -28,17 +28,17 @@ class LikeReviewAction
         $like = $this->likeRepository->findByCriteria($likeCriteria);
         $dislikeCriteria = new DislikeReviewCriteria($reviewId, $userId);
         $dislike = $this->dislikeRepository->findByCriteria($dislikeCriteria);
-        if ($like) {
-            $this->likeRepository->deleteById($like->id);
+        if ($dislike) {
+            $this->dislikeRepository->deleteById($dislike->id);
         }
-        if (empty($dislike)) {
-            $dislike = new Dislike([
-                'dislikeable_id' => $reviewId,
-                'dislikeable_type' => Review::class,
+        if (empty($like)) {
+            $like = new Like([
+                'likeable_id' => $reviewId,
+                'likeable_type' => Review::class,
                 'user_id' => $userId
             ]);
-            $this->dislikeRepository->save($dislike);
+            $this->likeRepository->save($like);
         }
-        return new DislikeReviewResponse();
+        return new LikeReviewResponse();
     }
 }
