+++ app/Actions/Like/LikeReviewResponse.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Hedonist\Actions\Like;
+
+class LikeReviewResponse 
+{
+
+}
