--- app/Actions/Review/Dislike/DislikeReviewResponse.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Hedonist\Actions\Review\Dislike;
-
-class DislikeReviewResponse 
-{
-
-}
