--- app/Actions/Review/Like/LikeReviewResponse.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Hedonist\Actions\Review\Like;
-
-class LikeReviewResponse 
-{
-
-}
